backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
alembic==1.13.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai==1.12.0
mcp==0.9.0